move reviewed submissions to reviewed tab
move reviewed submissions to reviewed tab
add tooltips
